refactor: extract bridge-transport and i18n seams out of desktop main.ts
Views (learning-content.ts, curriculum-wizard.ts) imported runBridge/t/tf
from the ~5200-line Tauri- and Three.js-coupled main.ts, which the MCP-Apps
panel cannot import without dragging in Tauri APIs and Three.js. Moves the
transport (now injectable via setBridgeTransport, with the Tauri invoke()
call installed from main.ts) and the i18n state/helpers (TRANSLATIONS,
currentLocale, t, tf) into two framework-free modules the panel can reuse
directly in the next task; main.ts re-exports runBridge/t/tf so any other
importer keeps working. Adds a guard test asserting the module boundaries
hold.
