+++ desktop/src/main.ts
@@ -6,791 +6,35 @@ import { open as openFolderDialog } from "@tauri-apps/plugin-dialog";
 import { openUrl } from "@tauri-apps/plugin-opener";
 import { check as checkForUpdate } from "@tauri-apps/plugin-updater";
 import * as THREE from "three";
+import { runBridge, setBridgeTransport } from "./bridge-transport.js";
 import {
   BLOOM_PACKS,
+  currentLocale,
   type Locale,
   LOCALE_LABELS,
   LOCALES,
   PRIVACY_PACKS,
-  TRANSLATION_PACKS,
+  setCurrentLocale,
+  t,
+  tf,
 } from "./i18n.js";
 import { initCurriculumWizard } from "./curriculum-wizard.js";
 import { initLearningContentStudio, loadStudioData } from "./learning-content.js";
 
-const ZAM_RELEASES_URL = "https://github.com/zam-os/zam/releases";
+// Re-exported so any other importer of "./main.js" keeps working unchanged;
+// learning-content.ts and curriculum-wizard.ts now import these directly
+// from bridge-transport.js/i18n.js instead.
+export { runBridge, t, tf };
 
-// ── LOCALIZATION DICTIONARIES ─────────────────────────────────────────────
-const TRANSLATIONS: Record<string, Record<string, string>> = {
-  en: {
-    ai_status_offline: "AI Offline",
-    ai_status_online: "Local AI Online",
-    ai_status_cloud_online: "Cloud AI · {model}",
-    ai_status_cloud_offline: "Cloud AI offline · {model}",
-    ai_status_starting: "Checking AI...",
-    ai_status_model_missing: "AI: model not found",
-    study_question_original: "Original",
-    study_evaluation_model: "{model}",
-    nav_dashboard: "Dashboard",
-    nav_settings: "Settings",
-    dashboard_kicker: "Today in ZAM",
-    dashboard_title: "Learn deliberately.",
-    dashboard_subtitle:
-      "Review what is due, open your knowledge map, or configure local AI and workspaces in Settings.",
-    lbl_due_reviews: "Due Reviews",
-    lbl_caught_up: "You're all caught up!",
-    dashboard_error: "Could not load your data",
-    lbl_domains: "Active Domains",
-    btn_start_session: "Start Learning Session",
-    lbl_translating: "Translating dynamically...",
-    placeholder_answer: "Type your conceptual answer here... (Ctrl+Enter to submit)",
-    btn_reveal_answer: "Submit & Reveal Answer",
-    lbl_ai_evaluating: "Reviewing your answer and generating feedback...",
-    lbl_ai_working: "(This may take a moment.)",
-    lbl_wait_warn: "⚠ Evaluation is taking longer than expected.",
-    btn_wait_keep: "Keep Waiting",
-    btn_wait_skip: "Skip Offline",
-    lbl_question_wait_warn:
-      "⚠ Generating this question is taking longer than expected.",
-    btn_question_use_saved: "Use Saved Question",
-    lbl_ai_feedback_title: "ZAM Feedback",
-    lbl_reveal_title: "Reference Answer",
-    lbl_rating_instruction: "Rate your active recall honestly:",
-    lbl_rate_1: "Again",
-    lbl_rate_2: "Hard",
-    lbl_rate_3: "Good",
-    lbl_rate_4: "Easy",
-    btn_pause_session: "Pause & Exit Session",
-    lbl_generating_question: "Generating dynamic question...",
-    token: "Token",
-    concept: "Concept",
-    context: "Context",
-    source: "Source Reference",
-    bloom_level: "Bloom Level",
-    rating_again_shortcut: "(Shortcut: 1)",
-    rating_hard_shortcut: "(Shortcut: 2)",
-    rating_good_shortcut: "(Shortcut: 3)",
-    rating_easy_shortcut: "(Shortcut: 4)",
-    session_completed: "Learning Session Completed!",
-    session_completed_sub: "Great job completing this session! Your memory traces have been updated.",
-    btn_back_to_dashboard: "Back to Dashboard",
-    btn_open_graph: "Knowledge Map (3D)",
-    btn_open_settings: "Settings",
-    observer_title: "UI Observer",
-    observer_idle: "Load windows and choose one application window to observe.",
-    observer_loading: "Loading observable windows...",
-    observer_ready: "Selected: {title}",
-    observer_vision_checking: "Checking vision observation settings...",
-    observer_analyzing: "Capturing snapshot and asking the vision model...",
-    observer_done: "Observation saved. Latest report confidence: {confidence}",
-    observer_canceled: "Observation canceled.",
-    observer_error: "Observer error: {message}",
-    observer_vision_disabled: "Vision observation is disabled. Enable it with: zam settings set llm.vision.enabled true",
-    observer_vision_offline: "Vision endpoint is offline: {url}",
-    observer_vision_model_missing: "Vision model is not available: {model}",
-    observer_privacy_paused: "Privacy pause: this window is blocked by the privacy filter ({reason}).",
-    observer_privacy_option: "privacy pause",
-    observer_refresh: "Load Windows",
-    observer_analyze: "Snapshot & Analyze",
-    observer_cancel: "Cancel",
-    observer_empty: "No observable windows found.",
-    observer_select_initial: "No windows loaded yet",
-    observer_select_placeholder: "Select a window",
-    observer_history_title: "Observation Reports",
-    observer_history_refresh: "Refresh Reports",
-    observer_history_empty: "No observation reports yet.",
-    observer_history_loaded: "Loaded {count} observation report(s).",
-    observer_loop_start: "Start Loop",
-    observer_loop_stop: "Stop Loop",
-    observer_loop_idle: "Manual snapshots only.",
-    observer_loop_running: "Observer loop running. One snapshot at a time.",
-    observer_loop_waiting: "Observer loop running. Next snapshot in {seconds}s.",
-    observer_loop_stopped: "Observer loop stopped.",
-    observer_watch_start: "Start Watch",
-    observer_watch_stop: "Stop Watch",
-    observer_watch_idle: "Continuous watch is off.",
-    observer_watch_starting: "Starting continuous watch...",
-    observer_watch_running: "Watching {title} — {count} event(s).",
-    observer_watch_stopping: "Stopping continuous watch...",
-    observer_watch_stopped: "Continuous watch stopped.",
-    observer_watch_error: "Watch error: {message}",
-    graph_title: "Knowledge Graph (3D)",
-    graph_hint: "Drag to rotate • Click nodes to focus • Scroll to zoom",
-    graph_focus: "Focus",
-    graph_prereqs: "Bases (Prerequisites)",
-    graph_dependents: "Higher Abilities (Dependents)",
-    graph_no_card: "no personal card yet",
-    graph_refresh: "Refresh",
-    graph_domain_token_list_title: "All tokens in this domain",
-    setup_title: "Setup & Data",
-    settings_kicker: "Local configuration",
-    settings_title: "Settings",
-    settings_subtitle:
-      "Configure this machine without forcing the same choices onto every workspace or device.",
-    settings_back: "Back to dashboard",
-    settings_system_title: "System",
-    settings_ai_title: "AI models",
-    settings_workspace_title: "Workspaces",
-    workspaces_help:
-      "Register personal, team, family, or community work directories for this machine.",
-    settings_appearance_title: "Appearance",
-    settings_data_title: "Data",
-    settings_database: "Database",
-    settings_learning_profile: "Learning profile",
-    database_checking: "Checking…",
-    database_status_local: "Local SQLite · connected",
-    database_status_turso: "Turso · connected",
-    database_status_error: "Connection unavailable",
-    database_detail: "{location} · profile {profile} · {count} cards",
-    database_no_profile: "No active profile",
-    database_profile_option: "{profile} — {count} cards",
-    database_profile_switch_confirm:
-      'Switch the active learning profile to "{profile}"?',
-    database_profile_switched: "Active profile: {profile} ({count} cards)",
-    database_refresh: "Refresh status",
-    settings_theme: "Theme",
-    theme_light: "Light",
-    theme_dark: "Dark",
-    btn_open_data_folder: "Open data folder",
-    btn_backup_db: "Back up database",
-    setup_backing_up: "Backing up database…",
-    setup_backed_up: "Backed up to {path}",
-    setup_remote_no_backup:
-      "Your database syncs to the cloud ({target}) — no separate local backup needed.",
-    setup_backup_failed: "Backup failed: {message}",
-    setup_backup_failed_generic: "Backup failed.",
-    setup_open_folder_failed: "Could not open the data folder: {message}",
-    lbl_workspace: "Workspace",
-    btn_choose_workspace: "+ Add workspace…",
-    btn_open_terminal: "Open Terminal",
-    terminal_opening: "Opening terminal...",
-    terminal_opened: "Opened terminal in {workspace}",
-    terminal_open_failed: "Could not open terminal: {message}",
-    workspace_active: "active",
-    workspace_default_label: "Default workspace",
-    workspace_empty: "No workspace registered yet.",
-    workspace_kind: "{kind} workspace",
-    workspace_kind_personal: "Personal workspace",
-    workspace_kind_team: "Team workspace",
-    workspace_kind_family: "Family workspace",
-    workspace_kind_community: "Community workspace",
-    workspace_kind_organization: "Organization workspace",
-    workspace_kind_custom: "Custom workspace",
-    workspace_more: "{count} more workspace(s) hidden. Use the CLI for the full list.",
-    workspace_use: "Use",
-    workspace_open: "Terminal",
-    workspace_default_suffix: "(default)",
-    workspace_set: "Workspace set to {path}",
-    workspace_added: "Workspace added: {path}",
-    workspace_pick_failed: "Could not set workspace: {message}",
-    workspace_remove: "Remove",
-    workspace_remove_confirm:
-      'Remove "{label}" from ZAM? Files and links in the folder will stay unchanged.',
-    workspace_removed: "Workspace removed: {label}",
-    workspace_remove_failed: "Could not remove workspace: {message}",
-    workspace_link_ok: "Skill link OK",
-    workspace_link_broken: "Skill link broken",
-    workspace_link_unmanaged: "Foreign skill folder",
-    workspace_repair: "Repair skill link",
-    workspace_repair_confirm:
-      'The folder for "{label}" contains a skills/zam directory that ZAM did not create. Repair deletes it and replaces it with a link. Continue?',
-    workspace_repairing: "Repairing skill link…",
-    workspace_repaired: "Skill link repaired: {label}",
-    workspace_repair_failed: "Could not repair skill link: {message}",
-    lbl_app_version: "Version",
-    lbl_learning_model: "Learning model",
-    lbl_observer_model: "Observer model",
-    provider_ready: "ready",
-    provider_disabled: "disabled",
-    provider_offline: "offline",
-    provider_model_missing: "model missing",
-    provider_unsupported: "unsupported provider",
-    provider_local: "local",
-    provider_cloud: "cloud",
-    provider_unknown: "unknown",
-    btn_ai_config_open: "Configure",
-    btn_ai_config_close: "Close",
-    btn_add_ai_provider: "+ Add provider",
-    ai_config_loading: "Loading provider configuration...",
-    ai_config_failed: "Could not load provider configuration: {message}",
-    ai_provider_empty: "No providers configured yet.",
-    ai_provider_section_local: "Local",
-    ai_provider_section_cloud: "Cloud",
-    ai_provider_kind: "Provider type",
-    ai_provider_kind_local: "Local",
-    ai_provider_kind_cloud: "Cloud",
-    ai_provider_local_badge: "local",
-    ai_provider_cloud_badge: "cloud",
-    ai_provider_key_set: "key set",
-    ai_provider_key_missing: "key missing",
-    ai_provider_key_none: "no key",
-    ai_provider_edit: "Edit",
-    ai_provider_remove: "Remove",
-    ai_provider_remove_confirm: 'Remove provider "{name}"?',
-    ai_provider_removed: 'Removed provider "{name}"',
-    ai_provider_saved: 'Saved provider "{name}"',
-    ai_provider_save_failed: "Could not save provider: {message}",
-    ai_provider_form_add_title: "Add provider",
-    ai_provider_form_edit_title: "Edit provider",
-    ai_provider_display_name: "Name",
-    ai_provider_url: "URL",
-    ai_provider_model: "Model",
-    ai_provider_model_local_hint:
-      "Type the model name. If the endpoint lists models, you can pick one below.",
-    ai_provider_models_pick: "Pick from endpoint",
-    ai_provider_flavor: "API flavor",
-    ai_provider_local: "Runs on this computer",
-    ai_provider_local_hint:
-      "Local model servers (FastFlowLM, Ollama, …) usually need no API key. ZAM can try to start the server automatically.",
-    ai_provider_cloud_hint_key:
-      "Cloud APIs need an API key. Local providers do not use the key field.",
-    ai_provider_runner: "Local model server",
-    ai_provider_runner_hint:
-      "The program that serves the model on this PC — not your coding agent.",
-    ai_provider_runner_auto: "Auto-detect",
-    ai_provider_runner_flm: "FastFlowLM",
-    ai_provider_runner_ollama: "Ollama",
-    ai_provider_runner_foundry: "Foundry Local",
-    ai_provider_runner_not_installed: "not detected",
-    ai_provider_api_key: "API key",
-    ai_provider_api_key_hint: "Write-only — stored locally, never synced",
-    ai_provider_cloud_hint: "Suggested model: {model}",
-    ai_provider_models_loading: "Loading models...",
-    ai_provider_models_empty: "No models listed by endpoint",
-    btn_ai_provider_save: "Save",
-    btn_ai_provider_cancel: "Cancel",
-    ai_role_bindings_title: "Role bindings",
-    ai_role_recall: "Learning (recall)",
-    ai_role_vision: "Observer (vision)",
-    ai_role_primary: "Primary",
-    ai_role_fallback: "Fallback",
-    ai_role_none: "(none)",
-    btn_ai_role_apply: "Apply",
-    ai_role_bound: "Role {role} updated",
-    ai_role_bind_failed: "Could not bind role: {message}",
-    ai_recall_anthropic_warn:
-      "Anthropic Messages providers are not supported for learning sessions yet. Bind anyway?",
-    ai_vision_cloud_confirm:
-      "Screenshots will be sent to {endpoint}. Enable vision observation and bind this cloud provider?",
-    ai_provider_referenced: "Still used by: {roles}",
-    btn_check_updates: "Check for updates",
-    btn_open_releases: "Releases",
-    version_unknown: "unknown",
-    update_checking: "Checking for updates...",
-    update_available: "Update available: {version}",
-    update_none: "You are on the latest version.",
-    update_failed: "Update check failed: {message}",
-    release_link_failed: "Could not open releases: {message}",
-    nav_content: "Learning Content",
-    content_title: "Learning Content Studio",
-    content_subtitle:
-      "Manage your personal learning cards, search tokens, and curate active recall prompts.",
-    btn_new_card: "New Card",
-    lbl_search_placeholder: "Search questions, answers, categories, keys...",
-    lbl_category_filter: "Category",
-    lbl_all_categories: "All Categories",
-    lbl_all_contexts: "All Contexts",
-    lbl_no_context_default: "No context default",
-    lbl_no_context_assignment: "No context assignment",
-    lbl_empty_content: "No learning cards found. Create your first card!",
-    lbl_empty_content_btn: "Create First Card",
-    lbl_question: "Question",
-    lbl_answer: "Answer / Concept",
-    lbl_category: "Category",
-    lbl_title: "Title (display name)",
-    lbl_source_link: "Source Link",
-    lbl_more_settings: "More Settings",
-    lbl_context: "Context",
-    lbl_bloom_level: "Bloom Level",
-    lbl_symbiosis_mode: "Symbiosis Mode",
-    lbl_slug: "Technical Key (Slug)",
-    lbl_slug_hint: "Automatically generated after creation",
-    btn_save: "Save Card",
-    btn_remove: "Remove Card",
-    btn_delete: "Hard Delete Token",
-    lbl_confirm_remove_title: "Confirm Card Removal",
-    lbl_confirm_remove_desc: "Removing this card will clear your personal FSRS learning state and history for this token. Other users or the global token catalog remain unaffected.",
-    lbl_confirm_delete_title: "Confirm Global Token Deletion",
-    lbl_confirm_delete_desc: "Permanently deleting this token will destroy the knowledge concept, the personal cards of all users, and all review history across the database.",
-    lbl_delete_impact: "This action affects:",
-    lbl_impact_cards: "{count} card(s)",
-    lbl_impact_reviews: "{count} review log(s)",
-    lbl_impact_steps: "{count} session step(s)",
-    lbl_impact_skills: "{count} agent skill(s) updated",
-    lbl_confirm_action: "Confirm",
-    lbl_cancel_action: "Cancel",
-    lbl_card_saved_toast: "Card saved successfully!",
-    lbl_card_removed_toast: "Card removed successfully!",
-    lbl_card_deleted_toast: "Token deleted successfully!",
-    lbl_error_saving: "Failed to save card",
-    lbl_error_loading: "Failed to load learning content",
-    lbl_card_status_new: "New",
-    lbl_card_status_learning: "Learning",
-    lbl_card_status_review: "Review",
-    lbl_card_status_relearning: "Relearning",
-    lbl_card_status_not_started: "Not Started",
-    lbl_card_due: "Due",
-    lbl_card_not_due: "Not Due",
-    btn_import_curriculum: "Import Curriculum",
-    lbl_import_modal_title: "Import Curriculum Standard",
-    lbl_import_text: "Curriculum Text",
-    placeholder_import_text: "Paste curriculum text or syllabus bullets here...",
-    lbl_import_source: "Source URL / Reference",
-    lbl_import_category: "Target Category",
-    lbl_import_progress_status: "Generating cards...",
-    lbl_import_progress_detail: "This may take up to a minute depending on your local LLM speed.",
-    btn_import_submit: "Import",
-    toast_import_success: "Successfully imported {createdCount} new tokens and ensured {ensuredCount} cards!",
-    lbl_error_importing: "Failed to import curriculum",
-    btn_curriculum_wizard: "Curriculum Wizard",
-    lbl_curriculum_wizard_title: "Curriculum Wizard",
-    wizard_step_country: "Country",
-    wizard_step_region: "Region",
-    wizard_step_schoolType: "School Type",
-    wizard_step_grade: "Grade",
-    wizard_step_subject: "Subject",
-    wizard_step_track: "Track",
-    wizard_step_topic: "Topics",
-    wizard_btn_back: "Back",
-    wizard_btn_next: "Next",
-    wizard_no_options: "No options available yet for this selection.",
-    wizard_err_select_option: "Please select an option to continue.",
-    wizard_err_no_topics: "Please select at least one topic.",
-    wizard_topic_scope_note: "Cards are generated from the complete curriculum page for this subject (all of its topics). Precise per-topic import is coming in a future update — for now, topic selection helps you track what to cover next.",
-    wizard_resume_prompt: "Continue where you left off:",
-    wizard_btn_resume: "Continue",
-    wizard_btn_restart: "Start over",
-    lbl_curriculum_wizard_loading: "Loading…",
-    wizard_hours: "{hours} hrs",
-    btn_split: "Split",
-    lbl_split_modal_title: "Split Card",
-    lbl_original_card_title: "Original Card (Summarized application question)",
-    lbl_split_original_question: "Question",
-    lbl_split_original_concept: "Answer / Concept",
-    lbl_split_action_title: "After split completed:",
-    lbl_action_block: "Keep original card and block it (requires new cards first)",
-    lbl_action_remove: "Remove original card",
-    lbl_split_progress_status: "Generating atomic proposals...",
-    lbl_split_progress_detail: "This may take up to a minute depending on your local LLM speed.",
-    lbl_atomic_proposals_title: "Atomic Card Proposals (Minimum 2)",
-    btn_split_modal_submit: "Confirm Split",
-    btn_content_foundations_card: "Foundations",
-    lbl_foundations_modal_title: "Import Foundations",
-    lbl_foundations_progress_status: "Generating foundational proposals...",
-    lbl_foundations_progress_detail: "This may take up to a minute depending on your local LLM speed.",
-    lbl_foundations_atomic_title: "Foundational Card Proposals",
-    btn_foundations_modal_submit: "Confirm Import",
-    lbl_import_source_type: "Source Type",
-    lbl_import_source_uri: "File Path or URL",
-    btn_import_source_analyze: "Analyze",
-    lbl_source_extracted_preview: "Extracted Plain Text Preview",
-    lbl_delete: "Delete",
-    lbl_proposal_number: "Proposal #{n}",
-    lbl_foundational_proposal_number: "Foundational Proposal #{n}",
-    lbl_foundation_existing_badge: "Existing card will be linked",
-    lbl_foundation_new_badge: "New card suggestion",
-    lbl_include: "Include",
-    lbl_err_analyze_source_first: "Please analyze a source file, web link, or OCR scan first.",
-    lbl_err_min_split_proposals: "At least 2 complete card proposals are required to split a card.",
-    lbl_err_select_foundation: "Please select at least one prerequisite proposal card to import.",
-    lbl_err_enter_path_or_url: "Please enter a file path or URL to analyze.",
-    lbl_analyzing_source: "Analyzing source content, please wait…",
-    lbl_err_analysis_failed: "Analysis failed",
-    lbl_err_analysis_prefix: "Analysis error",
-    lbl_err_concept_required: "Answer / concept is required.",
-    lbl_err_category_required: "Category is required.",
-    lbl_err_import_context_required: "Curriculum text is required.",
-    lbl_err_original_context_required: "Question and answer are required for the original card.",
-    settings_context_title: "Knowledge Context",
-    settings_context_help: "Select the default knowledge context for this device.",
-    settings_context_label: "Default Context",
-    wizard_context_label: "Assign imported cards to context:",
-  },
-  de: {
-    ai_status_offline: "KI offline",
-    ai_status_online: "Lokale KI online",
-    ai_status_cloud_online: "Cloud-KI · {model}",
-    ai_status_cloud_offline: "Cloud-KI offline · {model}",
-    ai_status_starting: "Prüfe KI...",
-    ai_status_model_missing: "KI: Modell fehlt",
-    study_question_original: "Urtext",
-    study_evaluation_model: "{model}",
-    nav_dashboard: "Übersicht",
-    nav_settings: "Einstellungen",
-    dashboard_kicker: "Heute in ZAM",
-    dashboard_title: "Bewusst lernen.",
-    dashboard_subtitle:
-      "Wiederhole, was fällig ist, öffne dein Wissensnetz oder konfiguriere lokale KI und Arbeitsbereiche in den Einstellungen.",
-    lbl_due_reviews: "Anstehende Wiederholungen",
-    lbl_caught_up: "Du bist voll auf dem Laufenden!",
-    dashboard_error: "Deine Daten konnten nicht geladen werden",
-    lbl_domains: "Aktive Wissensbereiche",
-    btn_start_session: "Lernsitzung starten",
-    lbl_translating: "Übersetze dynamisch...",
-    placeholder_answer: "Schreibe deine konzeptionelle Antwort... (Strg+Eingabe zum Absenden)",
-    btn_reveal_answer: "Antwort aufdecken & absenden",
-    lbl_ai_evaluating: "Deine Antwort wird geprüft und Feedback wird erstellt...",
-    lbl_ai_working: "(Das kann einen Moment dauern.)",
-    lbl_wait_warn: "⚠ Die Bewertung dauert ungewöhnlich lange...",
-    btn_wait_keep: "Weiter warten",
-    btn_wait_skip: "Offline fortfahren",
-    lbl_question_wait_warn:
-      "⚠ Das Erstellen der Frage dauert ungewöhnlich lange.",
-    btn_question_use_saved: "Gespeicherte Frage verwenden",
-    lbl_ai_feedback_title: "ZAM Feedback",
-    lbl_reveal_title: "Musterlösung",
-    lbl_rating_instruction: "Bewerte deine aktive Erinnerung ehrlich:",
-    lbl_rate_1: "Nochmal",
-    lbl_rate_2: "Schwer",
-    lbl_rate_3: "Gut",
-    lbl_rate_4: "Einfach",
-    btn_pause_session: "Pause & Sitzung beenden",
-    lbl_generating_question: "Erstelle dynamische Frage...",
-    token: "Token",
-    concept: "Konzept",
-    context: "Kontext",
-    source: "Quellen-Referenz",
-    bloom_level: "Bloom-Stufe",
-    rating_again_shortcut: "(Shortcut: 1)",
-    rating_hard_shortcut: "(Shortcut: 2)",
-    rating_good_shortcut: "(Shortcut: 3)",
-    rating_easy_shortcut: "(Shortcut: 4)",
-    session_completed: "Lernsitzung erfolgreich abgeschlossen!",
-    session_completed_sub: "Hervorragende Arbeit! Deine Gedächtnispfade wurden aktualisiert.",
-    btn_back_to_dashboard: "Zurück zur Übersicht",
-    btn_open_graph: "Wissensnetz (3D)",
-    btn_open_settings: "Einstellungen",
-    observer_title: "UI Observer",
-    observer_idle: "Fenster laden und ein Anwendungsfenster zur Beobachtung auswählen.",
-    observer_loading: "Beobachtbare Fenster werden geladen...",
-    observer_ready: "Ausgewählt: {title}",
-    observer_vision_checking: "Prüfe Vision-Beobachtungseinstellungen...",
-    observer_analyzing: "Erzeuge Snapshot und frage das Vision-Modell...",
-    observer_done: "Beobachtung gespeichert. Confidence des letzten Reports: {confidence}",
-    observer_canceled: "Beobachtung abgebrochen.",
-    observer_error: "Observer-Fehler: {message}",
-    observer_vision_disabled: "Vision-Beobachtung ist deaktiviert. Aktiviere sie mit: zam settings set llm.vision.enabled true",
-    observer_vision_offline: "Vision-Endpunkt ist offline: {url}",
-    observer_vision_model_missing: "Vision-Modell ist nicht verfügbar: {model}",
-    observer_privacy_paused: "Privacy-Pause: Dieses Fenster wird durch den Privacy-Filter blockiert ({reason}).",
-    observer_privacy_option: "Privacy-Pause",
-    observer_refresh: "Fenster laden",
-    observer_analyze: "Snapshot analysieren",
-    observer_cancel: "Abbrechen",
-    observer_empty: "Keine beobachtbaren Fenster gefunden.",
-    observer_select_initial: "Noch keine Fenster geladen",
-    observer_select_placeholder: "Fenster auswählen",
-    observer_history_title: "Beobachtungsberichte",
-    observer_history_refresh: "Berichte aktualisieren",
-    observer_history_empty: "Noch keine Beobachtungsberichte.",
-    observer_history_loaded: "{count} Beobachtungsbericht(e) geladen.",
-    observer_loop_start: "Loop starten",
-    observer_loop_stop: "Loop stoppen",
-    observer_loop_idle: "Nur manuelle Snapshots.",
-    observer_loop_running: "Observer-Loop läuft. Immer nur ein Snapshot gleichzeitig.",
-    observer_loop_waiting: "Observer-Loop läuft. Nächster Snapshot in {seconds}s.",
-    observer_loop_stopped: "Observer-Loop gestoppt.",
-    observer_watch_start: "Watch starten",
-    observer_watch_stop: "Watch stoppen",
-    observer_watch_idle: "Kontinuierliche Beobachtung ist aus.",
-    observer_watch_starting: "Kontinuierliche Beobachtung wird gestartet...",
-    observer_watch_running: "Beobachte {title} — {count} Ereignis(se).",
-    observer_watch_stopping: "Kontinuierliche Beobachtung wird gestoppt...",
-    observer_watch_stopped: "Kontinuierliche Beobachtung gestoppt.",
-    observer_watch_error: "Watch-Fehler: {message}",
-    graph_title: "Wissensnetz (3D)",
-    graph_hint: "Ziehen zum Drehen • Knoten klicken = Fokus • Scroll = Zoomen",
-    graph_focus: "Fokus",
-    graph_prereqs: "Basis (Voraussetzungen)",
-    graph_dependents: "Höhere Fähigkeiten (Darauf aufbauend)",
-    graph_no_card: "noch keine persönliche Karte",
-    graph_refresh: "Neu laden",
-    graph_domain_token_list_title: "Alle Tokens in diesem Bereich",
-    setup_title: "Einrichtung & Daten",
-    settings_kicker: "Lokale Konfiguration",
-    settings_title: "Einstellungen",
-    settings_subtitle:
-      "Konfiguriere diesen Rechner, ohne dieselben Entscheidungen auf jeden Arbeitsbereich oder jedes Gerät zu übertragen.",
-    settings_back: "Zurück zur Übersicht",
-    settings_system_title: "System",
-    settings_ai_title: "KI-Modelle",
-    settings_workspace_title: "Arbeitsbereiche",
-    workspaces_help:
-      "Registriere persönliche, Team-, Familien- oder Community-Arbeitsverzeichnisse für diesen Rechner.",
-    settings_appearance_title: "Darstellung",
-    settings_data_title: "Daten",
-    settings_database: "Datenbank",
-    settings_learning_profile: "Lernprofil",
-    database_checking: "Wird geprüft…",
-    database_status_local: "Lokales SQLite · verbunden",
-    database_status_turso: "Turso · verbunden",
-    database_status_error: "Verbindung nicht verfügbar",
-    database_detail: "{location} · Profil {profile} · {count} Karten",
-    database_no_profile: "Kein aktives Profil",
-    database_profile_option: "{profile} — {count} Karten",
-    database_profile_switch_confirm:
-      'Aktives Lernprofil zu „{profile}“ wechseln?',
-    database_profile_switched: "Aktives Profil: {profile} ({count} Karten)",
-    database_refresh: "Status aktualisieren",
-    settings_theme: "Theme",
-    theme_light: "Hell",
-    theme_dark: "Dunkel",
-    btn_open_data_folder: "Datenordner öffnen",
-    btn_backup_db: "Datenbank sichern",
-    setup_backing_up: "Datenbank wird gesichert…",
-    setup_backed_up: "Gesichert nach {path}",
-    setup_remote_no_backup:
-      "Deine Datenbank synchronisiert in die Cloud ({target}) — kein separates lokales Backup nötig.",
-    setup_backup_failed: "Sicherung fehlgeschlagen: {message}",
-    setup_backup_failed_generic: "Sicherung fehlgeschlagen.",
-    setup_open_folder_failed: "Datenordner konnte nicht geöffnet werden: {message}",
-    btn_choose_workspace: "+ Arbeitsbereich hinzufügen…",
-    btn_open_terminal: "Terminal öffnen",
-    terminal_opening: "Terminal wird geöffnet...",
-    terminal_opened: "Terminal in {workspace} geöffnet",
-    terminal_open_failed: "Terminal konnte nicht geöffnet werden: {message}",
-    workspace_active: "aktiv",
-    workspace_default_label: "Standard-Arbeitsbereich",
-    workspace_empty: "Noch kein Arbeitsbereich registriert.",
-    workspace_kind: "{kind}-Arbeitsbereich",
-    workspace_kind_personal: "Persönlicher Arbeitsbereich",
-    workspace_kind_team: "Team-Arbeitsbereich",
-    workspace_kind_family: "Familien-Arbeitsbereich",
-    workspace_kind_community: "Community-Arbeitsbereich",
-    workspace_kind_organization: "Organisations-Arbeitsbereich",
-    workspace_kind_custom: "Benutzerdefinierter Arbeitsbereich",
-    workspace_more: "{count} weitere Arbeitsbereiche ausgeblendet. Die vollständige Liste ist in der CLI.",
-    workspace_use: "Nutzen",
-    workspace_open: "Terminal",
-    workspace_default_suffix: "(Standard)",
-    workspace_set: "Arbeitsbereich gesetzt: {path}",
-    workspace_added: "Arbeitsbereich hinzugefügt: {path}",
-    workspace_pick_failed: "Arbeitsbereich konnte nicht gesetzt werden: {message}",
-    workspace_remove: "Entfernen",
-    workspace_remove_confirm:
-      '"{label}" aus ZAM entfernen? Dateien und Verknüpfungen im Ordner bleiben unverändert.',
-    workspace_removed: "Arbeitsbereich entfernt: {label}",
-    workspace_remove_failed: "Arbeitsbereich konnte nicht entfernt werden: {message}",
-    workspace_link_ok: "Skill-Link OK",
-    workspace_link_broken: "Skill-Link defekt",
-    workspace_link_unmanaged: "Fremder Skill-Ordner",
-    workspace_repair: "Skill-Link reparieren",
-    workspace_repair_confirm:
-      'Im Ordner von "{label}" liegt ein skills/zam-Verzeichnis, das nicht von ZAM stammt. Beim Reparieren wird es gelöscht und durch eine Verknüpfung ersetzt. Fortfahren?',
-    workspace_repairing: "Skill-Link wird repariert…",
-    workspace_repaired: "Skill-Link repariert: {label}",
-    workspace_repair_failed: "Skill-Link konnte nicht repariert werden: {message}",
-    lbl_app_version: "Version",
-    lbl_learning_model: "Lernmodell",
-    lbl_observer_model: "Observer-Modell",
-    provider_ready: "bereit",
-    provider_disabled: "deaktiviert",
-    provider_offline: "offline",
-    provider_model_missing: "Modell fehlt",
-    provider_unsupported: "Provider nicht unterstützt",
-    provider_local: "lokal",
-    provider_cloud: "Cloud",
-    provider_unknown: "unbekannt",
-    btn_ai_config_open: "Konfigurieren",
-    btn_ai_config_close: "Schließen",
-    btn_add_ai_provider: "+ Provider hinzufügen",
-    ai_config_loading: "Provider-Konfiguration wird geladen...",
-    ai_config_failed: "Provider-Konfiguration konnte nicht geladen werden: {message}",
-    ai_provider_empty: "Noch keine Provider konfiguriert.",
-    ai_provider_section_local: "Lokal",
-    ai_provider_section_cloud: "Cloud",
-    ai_provider_kind: "Provider-Typ",
-    ai_provider_kind_local: "Lokal",
-    ai_provider_kind_cloud: "Cloud",
-    ai_provider_local_badge: "lokal",
-    ai_provider_cloud_badge: "Cloud",
-    ai_provider_key_set: "Schlüssel gesetzt",
-    ai_provider_key_missing: "Schlüssel fehlt",
-    ai_provider_key_none: "kein Schlüssel",
-    ai_provider_edit: "Bearbeiten",
-    ai_provider_remove: "Entfernen",
-    ai_provider_remove_confirm: 'Provider "{name}" entfernen?',
-    ai_provider_removed: 'Provider "{name}" entfernt',
-    ai_provider_saved: 'Provider "{name}" gespeichert',
-    ai_provider_save_failed: "Provider konnte nicht gespeichert werden: {message}",
-    ai_provider_form_add_title: "Provider hinzufügen",
-    ai_provider_form_edit_title: "Provider bearbeiten",
-    ai_provider_display_name: "Name",
-    ai_provider_url: "URL",
-    ai_provider_model: "Modell",
-    ai_provider_model_local_hint:
-      "Modellnamen eintippen. Listet der Endpunkt Modelle, kannst du unten eines wählen.",
-    ai_provider_models_pick: "Vom Endpunkt übernehmen",
-    ai_provider_flavor: "API-Variante",
-    ai_provider_local: "Läuft auf diesem Rechner",
-    ai_provider_local_hint:
-      "Lokale Modell-Server (FastFlowLM, Ollama, …) brauchen meist keinen API-Schlüssel. ZAM kann den Server automatisch starten.",
-    ai_provider_cloud_hint_key:
-      "Cloud-APIs brauchen einen API-Schlüssel. Lokale Provider nutzen das Schlüsselfeld nicht.",
-    ai_provider_runner: "Lokaler Modell-Server",
-    ai_provider_runner_hint:
-      "Das Programm, das das Modell auf diesem PC bereitstellt — nicht der Coding-Agent.",
-    ai_provider_runner_auto: "Automatisch erkennen",
-    ai_provider_runner_flm: "FastFlowLM",
-    ai_provider_runner_ollama: "Ollama",
-    ai_provider_runner_foundry: "Foundry Local",
-    ai_provider_runner_not_installed: "nicht erkannt",
-    ai_provider_api_key: "API-Schlüssel",
-    ai_provider_api_key_hint: "Nur Schreiben — lokal gespeichert, nie synchronisiert",
-    ai_provider_cloud_hint: "Vorgeschlagenes Modell: {model}",
-    ai_provider_models_loading: "Modelle werden geladen...",
-    ai_provider_models_empty: "Endpunkt listet keine Modelle",
-    btn_ai_provider_save: "Speichern",
-    btn_ai_provider_cancel: "Abbrechen",
-    ai_role_bindings_title: "Rollen-Zuordnung",
-    ai_role_recall: "Lernen (recall)",
-    ai_role_vision: "Observer (vision)",
-    ai_role_primary: "Primär",
-    ai_role_fallback: "Fallback",
-    ai_role_none: "(keiner)",
-    btn_ai_role_apply: "Anwenden",
-    ai_role_bound: "Rolle {role} aktualisiert",
-    ai_role_bind_failed: "Rolle konnte nicht zugeordnet werden: {message}",
-    ai_recall_anthropic_warn:
-      "Anthropic-Messages-Provider werden für Lernsitzungen noch nicht unterstützt. Trotzdem zuordnen?",
-    ai_vision_cloud_confirm:
-      "Screenshots werden an {endpoint} gesendet. Vision-Beobachtung aktivieren und diesen Cloud-Provider zuordnen?",
-    ai_provider_referenced: "Noch verwendet von: {roles}",
-    btn_check_updates: "Nach Updates suchen",
-    btn_open_releases: "Releases",
-    version_unknown: "unbekannt",
-    update_checking: "Suche nach Updates...",
-    update_available: "Update verfügbar: {version}",
-    update_none: "Du nutzt die aktuelle Version.",
-    update_failed: "Update-Prüfung fehlgeschlagen: {message}",
-    release_link_failed: "Releases konnten nicht geöffnet werden: {message}",
-    nav_content: "Lerninhalte",
-    content_title: "Lerninhalt-Studio",
-    content_subtitle:
-      "Verwalte deine persönlichen Lernkarten, durchsuche Tokens und pflege aktive Erinnerungsfragen.",
-    btn_new_card: "Neue Karte",
-    lbl_search_placeholder: "Fragen, Antworten, Kategorien, Keys suchen...",
-    lbl_category_filter: "Kategorie",
-    lbl_all_categories: "Alle Kategorien",
-    lbl_all_contexts: "Alle Kontexte",
-    lbl_no_context_default: "Kein Kontextstandard",
-    lbl_no_context_assignment: "Keine Kontextzuordnung",
-    lbl_empty_content: "Keine Lernkarten gefunden. Erstelle deine erste karte!",
-    lbl_empty_content_btn: "Erste Karte erstellen",
-    lbl_question: "Frage",
-    lbl_answer: "Antwort / Konzept",
-    lbl_category: "Kategorie",
-    lbl_title: "Titel (Anzeigename)",
-    lbl_source_link: "Quell-Link",
-    lbl_more_settings: "Mehr Einstellungen",
-    lbl_context: "Kontext",
-    lbl_bloom_level: "Bloom-Level",
-    lbl_symbiosis_mode: "Symbiose-Modus",
-    lbl_slug: "Technischer Key (Slug)",
-    lbl_slug_hint: "Wird nach Erstellung automatisch generiert",
-    btn_save: "Karte speichern",
-    btn_remove: "Karte entfernen",
-    btn_delete: "Token dauerhaft löschen",
-    lbl_confirm_remove_title: "Kartenentfernung bestätigen",
-    lbl_confirm_remove_desc: "Das Entfernen dieser Karte löscht deinen persönlichen FSRS-Lernstatus und -Verlauf für diesen Token. Andere Benutzer oder der globale Token-Katalog bleiben unberührt.",
-    lbl_confirm_delete_title: "Globale Token-Löschung bestätigen",
-    lbl_confirm_delete_desc: "Das dauerhafte Löschen dieses Tokens zerstört das Wissenskonzept, die persönlichen Karten aller Benutzer und den gesamten Bewertungsverlauf in der Datenbank.",
-    lbl_delete_impact: "Diese Aktion betrifft:",
-    lbl_impact_cards: "{count} Karte(n)",
-    lbl_impact_reviews: "{count} Bewertungsprotokoll(e)",
-    lbl_impact_steps: "{count} Sessionschritt(e)",
-    lbl_impact_skills: "{count} Agenten-Skill(s) aktualisiert",
-    lbl_confirm_action: "Bestätigen",
-    lbl_cancel_action: "Abbrechen",
-    lbl_card_saved_toast: "Karte erfolgreich gespeichert!",
-    lbl_card_removed_toast: "Karte erfolgreich entfernt!",
-    lbl_card_deleted_toast: "Token erfolgreich gelöscht!",
-    lbl_error_saving: "Karte konnte nicht gespeichert werden",
-    lbl_error_loading: "Lerninhalte konnten nicht geladen werden",
-    lbl_card_status_new: "Neu",
-    lbl_card_status_learning: "Lernen",
-    lbl_card_status_review: "Wiederholen",
-    lbl_card_status_relearning: "Wiederlernen",
-    lbl_card_status_not_started: "Nicht gestartet",
-    lbl_card_due: "Fällig",
-    lbl_card_not_due: "Nicht fällig",
-    btn_import_curriculum: "Lehrplan importieren",
-    lbl_import_modal_title: "Lehrplan-Standard importieren",
-    lbl_import_text: "Lehrplantext",
-    placeholder_import_text: "Lehrplantext oder Lehrplanpunkte hier einfügen...",
-    lbl_import_source: "Quell-URL / Referenz",
-    lbl_import_category: "Zielkategorie",
-    lbl_import_progress_status: "Karten werden generiert...",
-    lbl_import_progress_detail: "Dies kann je nach Geschwindigkeit der lokalen KI bis zu einer Minute dauern.",
-    btn_import_submit: "Importieren",
-    toast_import_success: "Erfolgreich {createdCount} neue Token importiert und {ensuredCount} Karten sichergestellt!",
-    lbl_error_importing: "Lehrplan konnte nicht importiert werden",
-    btn_curriculum_wizard: "Lehrplan-Assistent",
-    lbl_curriculum_wizard_title: "Lehrplan-Assistent",
-    wizard_step_country: "Land",
-    wizard_step_region: "Bundesland",
-    wizard_step_schoolType: "Schulform",
-    wizard_step_grade: "Jahrgangsstufe",
-    wizard_step_subject: "Fach",
-    wizard_step_track: "Ausprägung",
-    wizard_step_topic: "Themen",
-    wizard_btn_back: "Zurück",
-    wizard_btn_next: "Weiter",
-    wizard_no_options: "Für diese Auswahl sind noch keine Optionen verfügbar.",
-    wizard_err_select_option: "Bitte wähle eine Option, um fortzufahren.",
-    wizard_err_no_topics: "Bitte wähle mindestens ein Thema aus.",
-    wizard_topic_scope_note: "Die Karten werden aus der vollständigen Lehrplanseite dieses Fachs erstellt (alle Themen). Die präzise Auswahl nur der markierten Themen kommt mit einem künftigen Update — bis dahin hilft dir die Auswahl, den Überblick zu behalten, was als Nächstes drankommt.",
-    wizard_resume_prompt: "Dort weitermachen, wo du aufgehört hast:",
-    wizard_btn_resume: "Fortfahren",
-    wizard_btn_restart: "Neu beginnen",
-    lbl_curriculum_wizard_loading: "Lädt…",
-    wizard_hours: "{hours} Std.",
-    btn_split: "Aufteilen",
-    lbl_split_modal_title: "Karte aufteilen",
-    lbl_original_card_title: "Originalkarte (Zusammenfassende Anwendungsfrage)",
-    lbl_split_original_question: "Frage",
-    lbl_split_original_concept: "Antwort / Konzept",
-    lbl_split_action_title: "Nach dem Aufteilen:",
-    lbl_action_block: "Originalkarte behalten und blockieren (erfordert zuerst neue Karten)",
-    lbl_action_remove: "Originalkarte entfernen",
-    lbl_split_progress_status: "Atomare Vorschläge werden generiert...",
-    lbl_split_progress_detail: "Dies kann je nach Geschwindigkeit der lokalen KI bis zu einer Minute dauern.",
-    lbl_atomic_proposals_title: "Atomare Kartenvorschläge (Mindestens 2)",
-    btn_split_modal_submit: "Aufteilung bestätigen",
-    btn_content_foundations_card: "Fundamente",
-    lbl_foundations_modal_title: "Fundamente importieren",
-    lbl_foundations_progress_status: "Fundamentvorschläge werden generiert...",
-    lbl_foundations_progress_detail: "Dies kann je nach Geschwindigkeit der lokalen KI bis zu einer Minute dauern.",
-    lbl_foundations_atomic_title: "Fundamentale Kartenvorschläge",
-    btn_foundations_modal_submit: "Import bestätigen",
-    lbl_import_source_type: "Quelltyp",
-    lbl_import_source_uri: "Dateipfad oder URL",
-    btn_import_source_analyze: "Analysieren",
-    lbl_source_extracted_preview: "Extrahierte Textvorschau",
-    lbl_delete: "Löschen",
-    lbl_proposal_number: "Vorschlag Nr. {n}",
-    lbl_foundational_proposal_number: "Grundlagen-Vorschlag Nr. {n}",
-    lbl_foundation_existing_badge: "Bestehende Karte wird verknüpft",
-    lbl_foundation_new_badge: "Neuer Kartenvorschlag",
-    lbl_include: "Einschließen",
-    lbl_err_analyze_source_first: "Bitte zuerst eine Quelldatei, einen Weblink oder einen OCR-Scan analysieren.",
-    lbl_err_min_split_proposals: "Für das Aufteilen einer Karte sind mindestens 2 vollständige Kartenvorschläge erforderlich.",
-    lbl_err_select_foundation: "Bitte wähle mindestens eine Grundlagen-Vorschlagskarte zum Importieren aus.",
-    lbl_err_enter_path_or_url: "Bitte gib einen Dateipfad oder eine URL zur Analyse ein.",
-    lbl_analyzing_source: "Quellinhalt wird analysiert, bitte warten…",
-    lbl_err_analysis_failed: "Analyse fehlgeschlagen",
-    lbl_err_analysis_prefix: "Analysefehler",
-    lbl_err_concept_required: "Antwort / Lerninhalt ist erforderlich.",
-    lbl_err_category_required: "Kategorie ist erforderlich.",
-    lbl_err_import_context_required: "Lehrplantext ist erforderlich.",
-    lbl_err_original_context_required: "Frage und Antwort sind für die Originalkarte erforderlich.",
-    settings_context_title: "Wissenskontext",
-    settings_context_help: "Wähle den Standard-Wissenskontext für dieses Gerät.",
-    settings_context_label: "Standard-Kontext",
-    wizard_context_label: "Importierte Karten dem Kontext zuweisen:",
-  },
-  // es, fr, pt, zh, ja live in ./i18n.ts; en/de stay here as reference locales.
-  ...TRANSLATION_PACKS,
-};
+// Installed at module scope — before any DOMContentLoaded handler can fire —
+// so every runBridge() call below (and in views that import it from
+// bridge-transport.js) reaches the Tauri backend starting with the first call.
+setBridgeTransport(async (cmd, args) => {
+  const raw = await invoke<string>("execute_zam_bridge", { cmd, args });
+  return JSON.parse(raw);
+});
+
+const ZAM_RELEASES_URL = "https://github.com/zam-os/zam/releases";
 
 const BLOOM_LEVEL_NAMES: Record<string, Record<number, string>> = {
   en: {
@@ -814,7 +58,6 @@ const BLOOM_LEVEL_NAMES: Record<string, Record<number, string>> = {
 type AppView = "dashboard-view" | "settings-view" | "study-view" | "graph-view" | "learning-content-view";
 type ThemePreference = "light" | "dark";
 
-let currentLocale = "en";
 let isLlmEnabled = false;
 let aiConfigEditorOpen = false;
 let aiConfigData: ProviderConfigListResponse | null = null;
@@ -1134,28 +377,6 @@ const OBSERVER_PRIVACY_REASON_LABELS: Record<string, Record<string, string>> = {
   },
   ...PRIVACY_PACKS,
 };
-
-// ── BRIDGE COMMAND RUNNER ────────────────────────────────────────────────
-export async function runBridge<T = any>(cmd: string, args: string[] = []): Promise<T> {
-  try {
-    const raw = await invoke<string>("execute_zam_bridge", { cmd, args });
-    return JSON.parse(raw) as T;
-  } catch (err) {
-    console.error(`Bridge Error [${cmd}]:`, err);
-    throw err;
-  }
-}
-
-export function t(key: string): string {
-  return TRANSLATIONS[currentLocale]?.[key] || TRANSLATIONS["en"]?.[key] || key;
-}
-
-export function tf(key: string, values: Record<string, string | number>): string {
-  return Object.entries(values).reduce(
-    (text, [name, value]) => text.split(`{${name}}`).join(String(value)),
-    t(key),
-  );
-}
 
 function loadThemePreference(): ThemePreference {
   try {
@@ -1606,7 +827,7 @@ function setupLocaleSwitcher(): void {
 
 async function setLocale(locale: Locale): Promise<void> {
   if (locale === currentLocale) return;
-  currentLocale = locale;
+  setCurrentLocale(locale);
 
   // Persist as an OS-default override; read back by desktop-bootstrap on launch.
   try {
@@ -4355,7 +3576,7 @@ async function loadDashboard() {
       workspaceDir?: string;
     }>("desktop-bootstrap");
     desktopUserId = settings.userId;
-    currentLocale = settings.locale || "en";
+    setCurrentLocale(settings.locale || "en");
     isLlmEnabled = settings.llm?.enabled || false;
     activeWorkspaceId = settings.activeWorkspaceId ?? activeWorkspaceId;
     activeWorkspaceDir = settings.workspaceDir ?? activeWorkspaceDir;
